programming

+++ include/defjams.h
@@ -142,7 +142,7 @@ typedef enum {
 typedef struct mixer_timing_info
 {
     int frames_per_midi_tick;
-    int midi_ticks_per_ms;
+    double midi_ticks_per_ms;
 
     double time_of_next_midi_tick;
